chore: move admin page to no footer
