+++ app/(noFooter)/admin/[email]/components/EventItem.tsx
