+++ app/(noFooter)/admin/[email]/components/StatisticsSection.tsx
