+++ app/(noFooter)/admin/[email]/components/types.ts
