+++ app/(noFooter)/admin/[email]/page.tsx
