+++ app/api/admin/stats/files/route.ts
@@ -1,7 +1,7 @@
 import { createClient } from '@/utils/supabase/server';
 import { Tables, Columns } from '@/schema/schema';
 import { NextResponse } from 'next/server';
-import { isAdminEmail } from '@/app/(withFooter)/admin/helpers';
+import { isAdminEmail } from '@/app/(noFooter)/admin/helpers';
 
 export async function GET() {
   const supabase = createClient();
