+++ app/api/admin/user-audit/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { createClient } from '@/utils/supabase/server';
 import { Tables, Columns } from '@/schema/schema';
-import { UserAuditData } from '@/app/(withFooter)/admin/[email]/components/types';
+import { UserAuditData } from '@/app/(noFooter)/admin/[email]/components/types';
 
 // Mark this route as dynamic to prevent static generation errors
 export const dynamic = 'force-dynamic';
